Add GSS code in the ONS code field as a test
Northern Irish authorities don't have ONS (aka SNAC) codes as the boundaries
were reorganised after the demise of the code format.  This means that some
of our services don't work well with NI, as we're heavily reliant on ONS
codes at present.
We're trying out a fix where we add the GSS code for the authorities in
where the ONS code would otherwise go.  This commit adds the GSS code for
Ards and North Down Borough Council as an initial test for Licensing.

https://trello.com/c/kjurG3i8/302-support-new-northern-ireland-local-authorities-for-licensing