Allow overriding of email subject prefix
In GOV.UK we want to specify different email subject lines for MapIt emails in
each environment.  This change allows the EMAIL_SUBJECT_PREFIX to be set in
general.yml.

The space after the closing ']' in the example is intentional in order to
format the full email subject nicely when it includes following text.