Update missing codes command to accept arguments
As part of the work to import local authority slugs, we need to be able to tell which Areas (that are Local Authorities) do not have a slug. We can reuse the existing 'mapit_UK_show_missing_codes' command by allowing it to accept arguments such as the `code_types` and `area_types`. We set the default values to be the previously hardcoded values so the script can still be called as it was before. However we do not want to check area_type 'EUR' for code_type 'govuk_slug', as we do not assign slugs to that area type. We add this exception to the code as a method call `is_code_type_required_for_area_type`.