Create command to import slugs for local authorities
MapIt will be the new canonical source for local authority slugs. We copy the authorities.json file from the Frontend app to MapIt. We also add the command to import the slugs from this file and assign them to the local authorities in MapIt.

We need to create a new CodeType called 'govuk_slug’. We step through each entry in the authorities.json file and retrieve the Area that matches the GSS code. For each matched Area, we create a new Code of CodeType 'govuk_slug' with the relevant slug.