Fix index out of range error for fix-2016-05 script
When getting the previous generation, the code was selecting the second element from the array. In the case of only having one Generation (as in our case when we drop the database before a new import), this will cause a `IndexError: list index out of range` error. The code change ensures that we have at least 2 elements in the array before picking the second element.