Update add_ons_to_gss command to support Python 2.
The command would error when running with Python 2.7, failing to
recognise the Welsh characters `ô` and `â` with "UnicodeEncodeError:
'ascii' codec can't encode character u'\xf4' in position 16: ordinal not
in range(128)". This is due to `csv.reader` changing between Python 2
and 3 - in 2 it expected a file-like object on which next() would return
raw bytes, and in 3 one which would return unicode strings. We opt to
explicitly check which version of Python is running and run the old code
for < Python 3.

We also stop assigning `name` for the 'code-change.csv' as it is not
used anywhere else in the script and the encoding probably won't be
handled correctly with Python 2. We stop assigning `type` and `name` for
'missing-codes.csv' for the same reasons.