Add DIS->CTY to the find parents command
District councils are all children of a county council, much like district
wards are children of a district council or county wards are children of a
county council.

One approach would be to write a script to explicitly connect all DIS
areas to the coorect CTY area, much like the old NI import scripts did
before we had boundary data (see:
mapit_gb/management/commands/mapit_UK_import_nspd_ni_areas.py and the
fixture mapit_gb/data/ni-electoral-areas.csv).

One downside to this is having to create a fixture of all the districts
and their correct county, which we'd need to update as boundaries change
and new councils are created or merged.  However, we know that (for now at
least) all districts live within a county so we can use geographic point
lookups to ask which CTY areas contain each DIS area.  The existing
mapit_UK_find_parents command already does this for other area types so
we just add DIS->CTY to the list of area types that it will consider.