Import names for local authorities
Local authority names from the boundary line import are not always the commonly
used ones.  We have a better source of the names with
[openregister's local authorities file](https://github.com/openregister/local-authority-data/blob/master/data/local-authority/local-authorities.tsv)

This file uses the partial ISO 3166:GB code as the identifier for local
authorities (missing the 'GB' country code), whereas we have GSS codes.  To
help us with mapping one dataset to another we used
[openregister's GSS mapping file](https://github.com/openregister/local-authority-data/blob/master/maps/gss.tsv)

We intend to use the register as the source of these names at some point in the
future, but we're happy to use this file for now.