Update location for the dump in S3
I've moved the mapit files into a directory in the same bucket today to keep
them more clearly separated from the other files in there, especially because
we're starting to mirror the source data there as well as keeping database
dumps. I also added the upload date to the start of the filenames so that they
sort more nicely. The file itself is unchanged so the hash doesn't change.