Fix long shebang line in pip
CI is failing with a message like this:
  `./test-venv/bin/pip install -r requirements.txt`
  `bash: ./test-venv/bin/pip: /var/lib/jenkins/workspace/t_update-for-feb-2017-onspd-SOR4L5PZIQPHZDGVUHAMTA: bad interpreter: No such file or directory`

When the virtualenv directory is `test-venv` then the shebang line at the top
of pip is 130 characters long which causes this error.  If we use `venv`
then we end up below the maximum of 127.

I've invoked pip using the venv python directly so that should paths change in
Jenkins in future, we should be OK.