Add OSM ref attribute as part of global import
This adds the `ref` attribute from OSM as a new kind of `Code` when
performing a global import.

This should make matching between Mapit/OSM and Wikidata simpler, which
in turn will help with the work we're doing in EveryPolitician at the
moment adding Mapit as a source of area data.