[UK] Use database for co-ordinate transformation.
If the application is running on a server with GDAL 1.9 (and maybe 1.8),
its OSGB co-ordinate transformation can be up to a few metres out. As we
use the transformation by the database when saving new coordinates, also
use it when fetching so we can match and display accurate information.

https://trac.osgeo.org/gdal/ticket/4597 is the GDAL ticket that explains
that it is fixed in GDAL 1.10.