Move configuration file to MapIt repo
This configuration file exists for integration, staging and production in
[alphagov-deployment][1].  We've now replaced the secrets held in the file with
environment variables and removed the bits that make the configuration
different.  This means that we can now use a generic `general.yml`, keep it
here and remove the old ones from alphagov-deployment.

Specifically, we've removed the rate limit configuration (as we allowed all
the things that would send traffic to MapIt anyway), and removed the error
email configuration as we now use Sentry.

[1]: https://github.com/alphagov/alphagov-deployment/tree/b3c79cead97f8fe927d96215a0eeada9aa12b7e7/mapit/settings