Add support for a 'country' filter in query parameters
This change means that you can add '?country=E' (say) to queries for
multiple areas to restrict the results to those areas that have a
Country with code 'E'.

This should be particularly helpful for global.mapit.mysociety.org
if we set an appropriate Country for each Area, since one is almost
always interested in results from a particular country.

Fixes #55