Improve performance of /nearest.
The nearest postcode lookup performance reduced dramatically when tried
with Postgres 9.6 and PostGIS 2.3. This commit updates the query to use
the ‘<->’ centroid distance operator instead, which makes better use of
the geoindex and is much faster to execute.

Handily, the work @dracos did to introduce the TrigramDistance function
into Django 1.10 was easily adaptable for this purpose, as the operator
used is the same.