Remove workaround for inefficient query planning
We have migrated MapIt to use Postgres 9.6 and PostGIS 2.3 now; Matthew
Somerville investigated whether this sub-optimal query planning still
happens with those versions, but it doesn't seem to. The query time is
considerably faster as well. So, we believe that this workaround should
no longer be necessary, but this commit should revert cleanly if in
practice we discover the problem is still present with some other
variant of these queries or data.