Update Mapit data
This data has been produced using the updated script in our [mapit-scripts repo](https://github.com/alphagov/mapit-scripts/pull/2).

Running the updated import script on each MapIt server will allow us to use
postcodes that were issued after August 2017.

An example of a postcode that didn't exist before this version is "B78 1LJ" so we should ensure that it is found otherwise there's no point in any of this.

To test this, you should run the updated `import-db-from-s3.sh` script from your VM and ensure that it runs to completion.

Then, fire up Mapit:
```
govuk_setenv mapit ./startup.sh
```
and check that we get a sensible response for some postcodes (this one is in Saffron Walden, in Uttlesford District Council).
```
curl http://localhost:3108/postcode/CB101HL
```

It's good to check some from around the UK, especially as the import for Northern Ireland is separate from other bits (BT1 1BG is in Belfast).