Manually prefetch any matching many-many countries
iterator() and prefetch() can't be used together, so the prefetch of many-many
countries is not happening in area lists, causing a slowdown if many areas are
to be returned. Instead, as we already do with codes, we manually fetch all of
the relevant countries from the join table and store them on the model objects
as we iterate through them.