Remove unnecessary python packages
We do not believe these are required.

`pygdal` is used to ease the installation of the python package `GDAL` in a virtualenv, but that package is no longer installed.

`numpy` was only added as a possible dependency of `pygdal`.

We do not see any evidence that `wsgiref` is in use.