Add note on naming conventions for downloaded dataset files
The `./import-uk-onspd` relies on a certain naming convention for dataset files.
As these files are imported from external sources we cannot guarantee their names.
This commit adds a note to the readme to ensure the folder we upload to S3 will
meet our needs.

This step could perhaps  be dealt with programmatically in the script instead, but
for now this is the path of least resistance.