Run under Python 3 in production
This changes the path of the virtual environment from `venv` to `venv3`
which is what it will be in production.

The reason why I've decided to give it a new directory name is that it's
not easy to upgrade a virtual environment in place.

This depends on
https://github.com/alphagov/govuk-app-deployment/pull/339 being merged
first and needs to be deployed with a hard restart.