Add some additional build dependencies and optional Python packages
This adds `libpq-dev`, needed to build `psycopg2`, and suggests in the
deployment script how to install `python-memcached` and `gdal` if they
are needed.

Although it may have been an option to include `python-memcached` in
`setup.py`, the Django `MemcachedCache` backend that makes use of this
is deprecated as of Django 3.2, so this will need to be replaced with
one of the supported backends.

See: https://github.com/mysociety/mapit/issues/390

Note that the GDAL install isn't included either as this must match the
version of GDAL on the system, hence the version check in the setup
script; assuming libgdal-dev is installed and the version matches, the
`pip install` should work on recent versions without setting
`C_INCLUDE_PATH` and `CPLUS_INCLUDE_PATH`.