Update handling of the find_module wrapper in utils.py
From Python >= 3.4 we're using importlib.machinery.PathFinder.find_spec
to determine whether there is a module for country-specific settings
available. This is change from Python < 3.1 where imp.find_module was
being used.

Rather than raising an exception if a module wasn't found, this returns
None so wasn't being caught by the try/except logic in settings.py,
resulting in the country module being added to INSTALLED_APPS regardless
of whether it could be found, causing a ModuleNotFoundError.

Python < 3.7 are no longer supported and we're currently testing against
Python 3.9 so this commit removes the version checks for earlier
versions when defining the `find_modules` wrapper and switches from a
try/except block in settings.py to a check for the result of the lookup
being None.